Refs codership/galera-bugs#915 Writeset serialization fails during IST
It likely to happen that IST donor locks seqno of already released action.
In fact it is very likely since only uncommitted actions are not released.
As of now `GCache::seqno_lock()` affects only results of top-level
`GCache::release()`, `GCache::discard_seqno()` and `GCache::discard_tail()`
calls and in the case of GCache::release() it is almost pointless as
mentioned above: a large part of "locked" actions already has been released.
At the same time underlying data stores (like the `RingBuffer` store) are
unaware of the locked seqno and may discard released actions which are
supposed to be protected by the lock in response to e.g. malloc() requests.
Pass locked seqno down to stores to inform them of discard limits.